meta: file indexer utility
Currently there is no index anywhere in the site, reaching for any
folder results in a 404 not found page. We suggested multiple times to
write an index.md file that just listes the articles present in the
folder of the same index.md file.

This utility is a python3 script that automaticall discover markdown
articles, gather metadata and update (or create) an index.md file at
the root of every folder.

See also prior work: #8