meta: improve text contrast
WebAIM (Accessibility in Mind) is a website that offer various stuff to
learn about and improve the accessibility of a page. [Contrast Checker]
is one of the their free tools, it rates the constrast of various
elements. With the current work we improved the rating from "red" (bad)
to "WCAG AA" on a scale "red" < WCAG AA < WCAG AAA.

Also the indentation in `layout.html` now uses tabs instead of spaces.

[Contrast Checker]: https://webaim.org/resources/contrastchecker/
Closes: #43