meta: new "commenter un article" issue template
We decided *not to* adopt a comment engine such as discourse for
discussing about the various article. We instead decided to extensively
use github issues in order to have a single place where all
contributions (including user contributions) would be.

This new template makes it clear that users are welcome to share their
feedback.

See also: #44