meta: renew C12G for templates and local jekyll
Since the original version of the `CONTRIBUTING.md` document, we decided to use issues templates to guide users into making better contribution/reporting, the new templates are also included in each article's footer for better integration.

In this work we renew our c12g document to remplace the lengthy explanations about "how to contribute to articles" by shorter sentences with a link to the relevant issue template.

We took the opportunity to also link the github tutorial about local jekyll installation and we fixed the broken link to the settings by a text breadcrumb