new: day to day commiting in git
Using git on a daly basic, it is often hard to always commit meaningful
commit messages. Often we just want to push a work in progress for the
sake of having a backup on GitHub. In this new tutorial we explain how
to best use git during the lifetime of a branch, whether it is during
initial development, PR preparation or review feedback.

Fixes: #67